@@ -4,65 +4,66 @@
   "description": "render dash components in react",
   "main": "dash_renderer/dash_renderer.min.js",
   "scripts": {
-    "lint": "./node_modules/.bin/eslint --quiet --fix .",
-    "lint:test": "./node_modules/.bin/eslint .",
+    "private::format.js-eslint": "eslint --quiet --fix .",
+    "private::format.js-prettier": "prettier --config .prettierrc --write \"src/**/*.js\"",
+    "private::lint.js-eslint": "eslint .",
+    "private::lint.js-prettier": "prettier --config .prettierrc \"src/**/*.js\" --list-different",
     "build:js": "webpack --build release",
     "build:dev": "webpack --build local",
     "build:local": "renderer build local",
     "build": "renderer build",
     "postbuild": "es-check es5 dash_renderer/*.js",
     "start": "webpack-serve ./webpack.serve.config.js",
-    "format": "prettier --config .prettierrc --write \"src/**/*.js\"",
-    "format:test": "prettier --config .prettierrc \"src/**/*.js\" --list-different",
-    "test": "npm run lint",
-    "test:js": "jest"
+    "test": "jest",
+    "format": "run-s private::format.*",
+    "lint": "run-s private::lint.*"
   },
   "author": "chriddyp",
   "license": "MIT",
   "dependencies": {
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
+    "@babel/polyfill": "7.8.7",
+    "cookie": "^0.4.0",
+    "dependency-graph": "^0.9.0",
     "prop-types": "15.7.2",
-    "cookie": "^0.3.1",
-    "dependency-graph": "^0.5.0",
     "fast-isnumeric": "^1.1.3",
-    "radium": "^0.22.1",
-    "ramda": "^0.26.1",
-    "react-redux": "^4.4.5",
-    "redux": "^3.4.0",
-    "redux-actions": "^0.9.1",
-    "redux-thunk": "^2.0.1",
-    "viz.js": "1.8.0"
+    "radium": "^0.26.0",
+    "ramda": "^0.27.0",
+    "react": "16.13.0",
+    "react-dom": "16.13.0",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5",
+    "redux-actions": "^2.6.5",
+    "redux-thunk": "^2.3.0",
+    "viz.js": "2.1.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.0",
-    "@babel/plugin-transform-modules-commonjs": "^7.6.0",
-    "@babel/polyfill": "7.7.0",
-    "@babel/preset-env": "^7.6.0",
-    "@babel/preset-react": "^7.0.0",
-    "@plotly/dash-component-plugins": "^1.0.2",
-    "@svgr/webpack": "^4.1.0",
-    "babel-eslint": "^10.0.3",
+    "@babel/core": "^7.8.7",
+    "@babel/plugin-transform-modules-commonjs": "^7.8.3",
+    "@babel/preset-env": "^7.8.7",
+    "@babel/preset-react": "^7.8.3",
+    "@plotly/dash-component-plugins": "^1.2.0",
+    "@svgr/webpack": "^5.2.0",
+    "babel-eslint": "^10.1.0",
     "babel-loader": "^8.0.6",
-    "css-loader": "^3.2.0",
-    "es-check": "^5.0.0",
-    "eslint": "^5.15.3",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^3.6.0",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-react": "^7.12.4",
-    "jest": "^24.9.0",
-    "prettier": "^1.16.4",
-    "prettier-eslint": "^8.8.2",
-    "prettier-eslint-cli": "^4.7.1",
+    "css-loader": "^3.4.2",
+    "es-check": "^5.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.18.3",
+    "jest": "^25.1.0",
+    "prettier": "^1.19.1",
+    "prettier-eslint": "^9.0.1",
+    "prettier-eslint-cli": "^5.0.0",
     "prettier-stylelint": "^0.4.2",
-    "style-loader": "^1.0.0",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.8",
-    "webpack-dev-server": "^3.1.11",
-    "webpack-serve": "^3.1.1",
-    "whatwg-fetch": "^2.0.2"
+    "style-loader": "^1.1.3",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
+    "webpack-serve": "^3.2.0",
+    "whatwg-fetch": "^3.0.0"
   }
 }
