@@ -75,10 +75,9 @@ export const storePrefix = '_dash_persistence.';
 function err(e) {
     const error = typeof e === 'string' ? new Error(e) : e;
 
-    /* eslint-disable no-console */
     // Send this to the console too, so it's still available with debug off
+    /* eslint-disable-next-line no-console */
     console.error(e);
-    /* eslint-disable no-console */
 
     return createAction('ON_ERROR')({
         myID: storePrefix,
@@ -319,7 +318,7 @@ export function recordUiEdit(layout, newProps, dispatch) {
 
     forEach(persistedProp => {
         const [propName, propPart] = persistedProp.split('.');
-        if (newProps[propName]) {
+        if (newProps[propName] !== undefined) {
             const storage = getStore(persistence_type, dispatch);
             const {extract} = getTransform(element, propName, propPart);
 
