@@ -9,12 +9,12 @@ def has_context(func):
     def assert_context(*args, **kwargs):
         if not flask.has_request_context():
             raise exceptions.MissingCallbackContextException(
-                (
-                    "dash.callback_context.{} "
-                    "is only available from a callback!"
-                ).format(getattr(func, "__name__"))
+                ("dash.callback_context.{} is only available from a callback!").format(
+                    getattr(func, "__name__")
+                )
             )
         return func(*args, **kwargs)
+
     return assert_context
 
 
