@@ -38,9 +38,7 @@ def format_tag(tag_name, attributes, inner="", closed=False, opened=False):
         tag += ">" + inner + "</{tag}>"
     return tag.format(
         tag=tag_name,
-        attributes=" ".join(
-            ['{}="{}"'.format(k, v) for k, v in attributes.items()]
-        ),
+        attributes=" ".join(['{}="{}"'.format(k, v) for k, v in attributes.items()]),
     )
 
 
@@ -61,47 +59,40 @@ def get_asset_path(requests_pathname, asset_path, asset_url_path):
 
 
 def get_relative_path(requests_pathname, path):
-    if requests_pathname == '/' and path == '':
-        return '/'
-    elif requests_pathname != '/' and path == '':
+    if requests_pathname == "/" and path == "":
+        return "/"
+    elif requests_pathname != "/" and path == "":
         return requests_pathname
-    elif not path.startswith('/'):
+    elif not path.startswith("/"):
         raise exceptions.UnsupportedRelativePath(
-            "Paths that aren't prefixed with a leading / are not supported.\n" +
-            "You supplied: {}".format(path)
+            "Paths that aren't prefixed with a leading / are not supported.\n"
+            + "You supplied: {}".format(path)
         )
-    return "/".join(
-        [
-            requests_pathname.rstrip("/"),
-            path.lstrip("/")
-        ]
-    )
+    return "/".join([requests_pathname.rstrip("/"), path.lstrip("/")])
 
 
 def strip_relative_path(requests_pathname, path):
     if path is None:
         return None
-    elif ((requests_pathname != '/' and
-            not path.startswith(requests_pathname.rstrip('/')))
-            or (requests_pathname == '/' and not path.startswith('/'))):
+    elif (
+        requests_pathname != "/" and not path.startswith(requests_pathname.rstrip("/"))
+    ) or (requests_pathname == "/" and not path.startswith("/")):
         raise exceptions.UnsupportedRelativePath(
-            "Paths that aren't prefixed with a leading " +
-            "requests_pathname_prefix are not supported.\n" +
-            "You supplied: {} and requests_pathname_prefix was {}".format(
-                path,
-                requests_pathname
+            "Paths that aren't prefixed with a leading "
+            + "requests_pathname_prefix are not supported.\n"
+            + "You supplied: {} and requests_pathname_prefix was {}".format(
+                path, requests_pathname
             )
         )
-    if (requests_pathname != '/' and
-            path.startswith(requests_pathname.rstrip('/'))):
+    if requests_pathname != "/" and path.startswith(requests_pathname.rstrip("/")):
         path = path.replace(
             # handle the case where the path might be `/my-dash-app`
             # but the requests_pathname_prefix is `/my-dash-app/`
-            requests_pathname.rstrip('/'),
-            '',
-            1
+            requests_pathname.rstrip("/"),
+            "",
+            1,
         )
-    return path.strip('/')
+    return path.strip("/")
 
 
 # pylint: disable=no-member
