@@ -89,37 +89,36 @@ class Component(with_metaclass(ComponentMeta, object)):
             # e.g. "The dash_core_components.Dropdown component (version 1.6.0)
             # with the ID "my-dropdown"
             try:
-                error_string_prefix = 'The `{}.{}` component (version {}){}'.format(
+                error_string_prefix = "The `{}.{}` component (version {}){}".format(
                     self._namespace,
                     self._type,
-                    getattr(__import__(self._namespace), '__version__', 'unknown'),
-                    ' with the ID "{}"'.format(kwargs['id'])
-                    if 'id' in kwargs else ''
+                    getattr(__import__(self._namespace), "__version__", "unknown"),
+                    ' with the ID "{}"'.format(kwargs["id"]) if "id" in kwargs else "",
                 )
             except ImportError:
                 # Our tests create mock components with libraries that
                 # aren't importable
-                error_string_prefix = 'The `{}` component{}'.format(
+                error_string_prefix = "The `{}` component{}".format(
                     self._type,
-                    ' with the ID "{}"'.format(kwargs['id'])
-                    if 'id' in kwargs else ''
+                    ' with the ID "{}"'.format(kwargs["id"]) if "id" in kwargs else "",
                 )
 
             if not k_in_propnames and not k_in_wildcards:
                 raise TypeError(
                     "{} received an unexpected keyword argument: `{}`".format(
                         error_string_prefix, k
-                    ) + "\nAllowed arguments: {}".format(  # pylint: disable=no-member
+                    )
+                    + "\nAllowed arguments: {}".format(  # pylint: disable=no-member
                         ", ".join(sorted(self._prop_names))
                     )
                 )
 
             if k != "children" and isinstance(v, Component):
                 raise TypeError(
-                    error_string_prefix +
-                    " detected a Component for a prop other than `children`\n" +
-                    "Did you forget to wrap multiple `children` in an array?\n" +
-                    "Prop {} has value {}\n".format(k, repr(v))
+                    error_string_prefix
+                    + " detected a Component for a prop other than `children`\n"
+                    + "Did you forget to wrap multiple `children` in an array?\n"
+                    + "Prop {} has value {}\n".format(k, repr(v))
                 )
 
             if k == "id":
@@ -127,13 +126,13 @@ class Component(with_metaclass(ComponentMeta, object)):
                     for id_key, id_val in v.items():
                         if not isinstance(id_key, _strings):
                             raise TypeError(
-                                "dict id keys must be strings,\n" +
-                                "found {!r} in id {!r}".format(id_key, v)
+                                "dict id keys must be strings,\n"
+                                + "found {!r} in id {!r}".format(id_key, v)
                             )
                         if not isinstance(id_val, _strings + (int, float, bool)):
                             raise TypeError(
-                                "dict id values must be strings, numbers or bools,\n" +
-                                "found {!r} in id {!r}".format(id_val, v)
+                                "dict id values must be strings, numbers or bools,\n"
+                                + "found {!r} in id {!r}".format(id_val, v)
                             )
                 elif not isinstance(v, _strings):
                     raise TypeError(
@@ -331,10 +330,7 @@ class Component(with_metaclass(ComponentMeta, object)):
         ] + [
             c
             for c in self.__dict__
-            if any(
-                c.startswith(wc_attr)
-                for wc_attr in self._valid_wildcard_attributes
-            )
+            if any(c.startswith(wc_attr) for wc_attr in self._valid_wildcard_attributes)
         ]
         if any(p != "children" for p in props_with_values):
             props_string = ", ".join(
