@@ -22,15 +22,15 @@ DELAY_TIME = 0.2
 def create_race_conditions_test(endpoints):
     def test(self):
         app = Dash()
-        app.layout = html.Div([
-            html.Div('Hello world', id='output'),
-            dcc.Input(id='input', value='initial value')
-        ])
+        app.layout = html.Div(
+            [
+                html.Div("Hello world", id="output"),
+                dcc.Input(id="input", value="initial value"),
+            ]
+        )
         app.scripts.config.serve_locally = True
 
-        @app.callback(
-            Output('output', 'children'),
-            [Input('input', 'value')])
+        @app.callback(Output("output", "children"), [Input("input", "value")])
         def update(value):
             return value
 
@@ -42,8 +42,8 @@ def create_race_conditions_test(endpoints):
         def element_text(id):
             try:
                 return self.driver.find_element_by_id(id).text
-            except Exception:
-                return ''
+            except:
+                return ""
 
         app.server.before_request(delay)
         self.startServer(app)
@@ -54,11 +54,10 @@ def create_race_conditions_test(endpoints):
         time.sleep(total_delay + DELAY_TIME)
 
         wait_for(
-            lambda: element_text('output') == 'initial value',
+            lambda: element_text("output") == "initial value",
             lambda: '"{}" != "initial value"\nbody text: {}'.format(
-                element_text('output'),
-                element_text('react-entry-point')
-            )
+                element_text("output"), element_text("react-entry-point")
+            ),
         )
 
         self.assertTrue(self.is_console_clean())
@@ -67,10 +66,10 @@ def create_race_conditions_test(endpoints):
 
 
 routes = [
-    'layout',
-    'dependencies',
-    'update-component',
-    '_config'
+    "layout",
+    "dependencies",
+    "update-component",
+    "_config"
     # routes and component-suites
     # are other endpoints but are excluded to speed up tests
 ]
@@ -78,9 +77,6 @@ routes = [
 for route_list in itertools.permutations(routes, len(routes)):
     setattr(
         Tests,
-        'test_delayed_{}'.format(
-            '_'.join([
-                r.replace('-', '_') for r in route_list
-            ])),
-        create_race_conditions_test(route_list)
+        "test_delayed_{}".format("_".join([r.replace("-", "_") for r in route_list])),
+        create_race_conditions_test(route_list),
     )
